+++ NEWS.md
@@ -1,10 +1,11 @@
 # v 0.1.2
 
-* switched to a pkgdown site with gh actions
-* switched to R CMD check with gh actions
-* switch testing to `tinytest`
+* switched to a pkgdown site with gh actions [#1]
+* switched to R CMD check with gh actions [#2]
+* switch testing to `tinytest` [#3]
 * updated links throughout, made a redirect page for old site
 * fix minor bug to allow flexible column names in `filter_qa` [#4], `model_params` [#6]
+* added testing for flexible column names [#5] [#7]
 
 # v 0.1.1 (2019-02-13) 
 * fixed data.table recycling change ([#23](https://gitlab.com/robit.a/irg/issues/23))